Implement Error Handling

+++ TaskTrackingAPI/TaskTrackingAPI/Controllers/TaskDetailsController.cs
@@ -25,37 +25,81 @@ namespace TaskTrackingAPI.Controllers
         [HttpGet]
         public async Task<ActionResult<IEnumerable<TaskDetails>>> GetTaskDetails()
         {
-            return (await _taskRepository.GetTaskDetails()).ToList();
+            try
+            {
+                var tasks = (await _taskRepository.GetTaskDetails()).ToList();
+                if (tasks == null)
+                {
+                    return NotFound();
+                }
+
+                return tasks;
+            }
+            catch (Exception)
+            {
+                return BadRequest();
+            }
+
         }
 
         // GET: api/TaskDetails/5
         [HttpGet("{id}")]
-        public async Task<ActionResult<TaskDetails>> GetTaskDetails(long id)
+        public async Task<ActionResult<TaskDetails>> GetTaskDetails(int? id)
         {
-            var taskDetails = await _taskRepository.GetTaskDetail(id);
-
-            if (taskDetails == null)
+            if (id == null)
             {
-                return NotFound();
+                return BadRequest();
             }
 
-            return taskDetails;
+            try
+            {
+                var taskDetails = await _taskRepository.GetTaskDetail(id);
+
+                if (taskDetails == null)
+                {
+                    return NotFound();
+                }
+
+                return taskDetails;
+            }
+            catch (Exception)
+            {
+                return BadRequest();
+            }
+
         }
 
         // PUT: api/TaskDetails/5
         // To protect from overposting attacks, enable the specific properties you want to bind to, for
         // more details, see https://go.microsoft.com/fwlink/?linkid=2123754.
         [HttpPut("{id}")]
-        public async Task<IActionResult> PutTaskDetails(long id, TaskDetails taskDetails)
+        public async Task<IActionResult> PutTaskDetails(int id, TaskDetails taskDetails)
         {
             if (id != taskDetails.Id)
             {
                 return BadRequest();
             }
 
-            await _taskRepository.UpdateTaskDetail(taskDetails);
+            if (ModelState.IsValid)
+            {
+                try
+                {
+                    await _taskRepository.UpdateTaskDetail(taskDetails);
 
-            return NoContent();
+                    return Ok();
+                }
+                catch (Exception ex)
+                {
+                    if (ex.GetType().FullName == "Microsoft.EntityFrameworkCore.DbUpdateConcurrencyException")
+                    {
+                        return NotFound();
+                    }
+
+                    return BadRequest();
+                }
+            }
+
+            return BadRequest();
         }
 
         // POST: api/TaskDetails
@@ -64,25 +108,57 @@ namespace TaskTrackingAPI.Controllers
         [HttpPost]
         public async Task<ActionResult<TaskDetails>> PostTaskDetails(TaskDetails taskDetails)
         {
-            await _taskRepository.AddTaskDetail(taskDetails);
+            if (ModelState.IsValid)
+            {
 
-            return CreatedAtAction("GetTaskDetails", new { id = taskDetails.Id }, taskDetails);
+                try
+                {
+                    await _taskRepository.AddTaskDetail(taskDetails);
+                    if (taskDetails.Id > 0)
+                    {
+                        return CreatedAtAction("GetTaskDetails", new { id = taskDetails.Id }, taskDetails);
+
+                    }
+                    else
+                    {
+                        return NotFound();
+                    }
+                }
+                catch (Exception)
+                {
+                    return BadRequest();
+                }
+            }
+            return BadRequest();
         }
 
         // DELETE: api/TaskDetails/5
         [HttpDelete("{id}")]
-        public async Task<ActionResult<TaskDetails>> DeleteTaskDetails(long id)
+        public async Task<ActionResult<TaskDetails>> DeleteTaskDetails(int? id)
         {
-            var model = await _taskRepository.GetTaskDetail(id);
-
-            if (model == null)
+            if (id == null)
             {
-                return NotFound();
+                return BadRequest();
             }
 
-            await _taskRepository.DeleteTaskDetail(id);
+            try
+            {
+                var model = await _taskRepository.GetTaskDetail(id);
 
-            return model;
+                if (model == null)
+                {
+                    return NotFound();
+                }
+
+                await _taskRepository.DeleteTaskDetail(id);
+
+                return model;
+
+            }
+            catch (Exception)
+            {
+                return BadRequest();
+            }
         }
     }
 }
